summarization/requirements.txt
PyPDF2>=3.0.0
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.0.0
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
